+++ CHANGELOG.md
@@ -90,3 +90,4 @@ January 15th, 2024
     - The SystemD service page is now reached from a dedicated "management" utilities page.
     - Added an "advanced mode" configuration value to enable and disable potentially destructive administration tools.
 - Added Predator state indicator icons to dashboard.
+- Added working directory and interface directory file viewer.